More robust headroom calculation for device
When calculating headroom we used the following for device headroom:
- current consumption
- expected consumption when set by a flow
- 1kW

Now we use the max of current consumption and expected consumption.
Reason being that e.g a heater where we expect 3kW may still decice to
use only one element resulting in 1kW actual draw. For headroom
calculation, we still want to use 3kW though as that is it's current
setting.